ai_image_search/pyproject.toml
[project]
name = "ai-image-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.20.0",
    "dotenv>=0.9.9",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
]
